mouse: skip button mappings only for core devices.
Since X server 1.4, devices are unlikely to be marked as IsXExtensionDevice.
More common is IsXExtensionKeyboard or IsXExtensionPointer. However, with
many pointer devices also having keys (e.g. most wireless combos) a check
for IsXExtensionPointer is unreliable.

In gsd-mouse-manager, skip the XInput button mapping only if the device is a
core device. Otherwise, test for buttons.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>