mouse: allow left-handed setting for touchpads.
Button mapping is handled in the server, tapping is handled in the driver.
Thus, if the tapping is enabled and the device is left-handed, the tapping
order must be reversed to counter the effect of the left-handed button
mapping.

For right-handed setup:
phys. button 1          → logical button 1
one-finger tapping      → phys button 1

For left-handed setup:
phys. button 3          → logical button 1
one-finger-tapping      → phys. button 3

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>