Handle the XF86RotateWindows hotkey by rotating a laptop's display
This is for tablets that have a "rotate the screen" hotkey.  We find
the laptop's built-in video output and rotate it to the next available
rotation.

Also, don't assume that the configuration and the rr_screen come from
the GxdXrandrManager object.  That is the case only while the tray
icon's popup menu is active, but not when we are just handling the
XF86RotateWindows hotkey.

Signed-off-by: Federico Mena Quintero <federico@novell.com>