Start an on-screen-display window (OSD)
This is the code from GsdMediaKeysWindow without the parts specific to media-keys;
the basic GsdOsdWindow will just know how to draw and composite itself, but it will
have no actual contents.

The idea is to have subclasses in the media-keys plugin and in the RANDR plugin;
those can add their own contents to the window.

Signed-off-by: Federico Mena Quintero <federico@novell.com>