housekeeping: use a notification for low space warning
Only fallback to a dialog if the notification server doesn't
support actions.

Further work is needed to provide a way to blacklist volumes.  Also
we probably don't want to run these notifications in a loop over
volumes since notifications are async and it isn't super nice to
send a stream of bubbles to the user anyway.  Better to say
multiple volumes are low on space.