Add wacom g-s-d plugin.
This plugin is essentially a copy of the mouse plugin with a lot of
search/replace for the framework needed (hence the copyright messages).

Since the wacom driver puts a lot of emphasis on the different tools, the
schemas create a nested structure. The keys are stored in
org.gnome.settings-daemon.peripherals.wacom.{stylus|erasor|cursor|pad},
with tablet-wide settings in wacom and the tool-specific settings in the
tool-specific schema.

Supported keys with this commit:
.wacom rotation .. for tablet rotation in 90deg steps
.wacom touch    .. only emit events when touching the device
.wacom tablet-pc-button .. tablet PC stylus button behaviour

.wacom.stylus pressureshreshold .. threshold before click is registered
.wacom.stylus pressurecurve	.. adjust 'hardness/softness' of the tip
.wacom.stylus is-absolute	.. switch between relative/absolute mode
.wacom.stylus area		.. usable area (also for calibration)
.wacom.stylus buttonmapping	.. logical button mapping

The above settings are also available for eraser, and subsets on
cursor and pad.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>