updates: if we failed to get the updates 10 times in a row, then show the user a notification
This is a compomise we've come up with in #gnome-design where we need to
keep the user informed if they are not getting updates done, but also
we don't want to bombard the user with notifications like we have done
in the past.

By default, we only do the updates check once per day, so this shouldn't
lead to lots of extra messages.