common: Launch a custom script on input devices
A number of users require input configurations that go beyond what the
gnome-control-center and gnome-settings-daemon offer. Some of these
configurations are readily accessible through various commandline
interfaces (synclient, xinput, xsetwacom, etc.) but are too uncommon to
justify full inclusion into g-s-d and/or g-c-c.

This patch adds a new key
  org.gnome.settings-daemon.peripherals.input-devices hotplug-command

that lets the user specify a custom script to be executed whenever
devices are added or removed. The script is called with three arguments:

  path/to/custom/command -t [added|removed|present] -i <device ID> <device name>

The first argument specifies if the device was just added or removed
or is already present at g-s-d startup time. The device name is the name
as provided by XInput.

The custom command is called for all slave devices (incl. floating ones)
but not for master devices.

Note that the functionality is not currently called for any input
devices.

https://bugzilla.gnome.org/show_bug.cgi?id=635486