color: Register icc profile files with colord
Watch the user directories where programs may install ICC files and register
them with colord when they appear and unregister them from colord when they
are deleted.

This is required as colord is unable to scan the users home directory, for
three principal reasons:

1. Layering design violation
2. SELinux won't let a daemon running as root grub around in the users $HOME.
3. The home directory may be encrypted until after login.

This code is simply moved from the old gcm-session process. An optional lcms2
dependancy is added, which allows us to get the profile embedded checksum
without reading all the bytes and running it through an MD5 transform.

For a moderate number of large CMYK profiles (~5Mb) it's actually quite slow to
do this, but most users will just have the tiny EDID profiles which are really
quick to parse. If you've got GIMP or G-C-M installed, you already have lcms2
installed, and as a library it's _tiny_. So it's recommended.

As an added safety measure, the plugin will not recurse more than 2 levels deep
when searching for profiles.