orientation: Use new orientation property from udev
Instead of needing to poke at X devices ourselves.

Now we just read the new orientation from the
ID_INPUT_ACCELEROMETER_ORIENTATION property, and rotate the
screen if it's changed from the previous value.

This requires a newer version of udev, currently only in git,
but shortly in version 172. See also:
http://git.kernel.org/?p=linux/hotplug/udev.git;a=commit;h=24569e24dc94a7cffb8031eb0055e8d06cbdcb72

https://bugzilla.gnome.org/show_bug.cgi?id=650603