color: Remove the hacky heuristic to ensure that the correct profiles are checked for age
This removes the heuristic that assumed the name of any profiles prefixed with
'GCM' were created by GNOME Color Manager in response to a user calibration.

Note: This doesn't actually change any behaviour; the user only gets
notifications that the profile is out of calibration if the gsettings keys is
enabled manually.