clipboard: fix incremental sending from the clipboard manager
When the data to send from the clipboard is bigger than the maximum
allowed X selection size, we set the property value to XA_INCR, to
signal that the data should be sent in incremental chunks.
Unfortunately this never worked, because the manager didn't listen to
the property changes on the requestor window, so it didn't know when to
start sending data.

https://bugzilla.gnome.org/show_bug.cgi?id=652609