power: Add the idle actions
This ports the functionality from gnome-power-manager to dim the screen on idle,
and then power down the screen (using DPMS) and then finally to put the machine
to sleep.

In doing this, we watch the inhibited state of gnome-session to make sure we
do the right thing when the user is watching a video and that kind of thing.