power: Remove some unused schema for enabling the sleep-inactive actions
This was originally added in gnome-3-2 as the designers specified having a
checkbox next to the sleep timeout to control whether the action was enabled or
disabled, and we didn't want to reset the action to the default every time the
user clicked and unclicked the enable button.

It turns out the designers changed thier mind on the design, and so it was
changed late in the GNOME 3.2 cycle. These obsolete config options remained.

Resolves half of https://bugzilla.gnome.org/show_bug.cgi?id=660395