color: Do not load all the color devices twice at startup
If the color plugin was the first thing to use GnomeSettingsSession then we'll
get an unknown->active transision when starting up, and we'll try to re-assign
all the devices again.

This speeds up loading of the color plugin typically by ~250ms, or quite a bit
more if you have lots of devices like me.