Monitor the laptop's lid state
If the lid is closed, we disable the laptop's output.  If it gets open, we restore
a saved configuration or auto-configure the outputs.

We don't turn off the laptop's display, even if the lid is closed, if that is the
only display available.  This should aid debugging even with broken hardware that
does not report the lid's state correctly.

Signed-off-by: Federico Mena Quintero <federico@gnome.org>