wacom: Add functions to locate tablet display
Adds functions to get information about the display a tablet
should be mapped to. The "display" key in gsettings is first
consulted for EDID data and outputs scanned for a match. If
none is found, we look for one heuristically (in this case,
find a screen from vendor "WAC"). If still nothing is found,
we assume whole desktop (well, whole "screen") mapping.

https://bugzilla.gnome.org/show_bug.cgi?id=657423