color: Set the internal panel brightness if specified in the profile
Hardcore color geeks know that a color profile is only valid for the same
display luminance that was used during the calibration. If a user has gone to
the trouble of adding the extra metadata, then assume they want this specific
panel brightness set for a given profile.

This should keep the studios happy, as different luminance values should be used
by different industry defined spaces. When we know the display luminance for
the display percentage, we can ask the profile vendors to include this in thier
profiles as extra metadata. Until then, we can use the panel-specific percentage
values to be used as an effective workaround.

NOTE: We only support changing the internal panel, as this is all that the
gnome-settings-daemon interface controls. Even if the DBus interface was
cleverer, we've got no way of telling X to adjust the brightness of external
panels just yet.