xrandr: Fix the rotate display button not working
When hitting the XF86RotateWindows the screen just goes black instead of
actually rotating the screen as intending (with the output turned off).

Turns out that gnome-settings-daemon is trying to rotate the screen to
an invalid value. GnomeRRRotation is an enum with only values >= 0, so
it is likely to be made unsigned. As the rotation argument to
handle_rotate_windows can be -1 to indicate ``next possible rotation'',
rotation can't be a GnomeRRRotation as that won't be < 0.

https://bugzilla.gnome.org/show_bug.cgi?id=664363