bgo#657201 [power] - Allow forcing suspend when closing the lid in the presence of external monitors
Some people want, "suspend the laptop when I close its lid, even if an external monitor
is present", so that they can close the lid at the end of the day.  Others want,
"don't suspend the laptop when I close its lid if an external monitor is present",
which is the behavior introduced in bgo#663126, and serves common usage of docking stations
or laptops as media players when connected to a TV.  This commit introduces
an org.gnome.settings-daemon.plugins.power.lid-close-suspend-with-external-monitors key,
with a default of False, that allows changing the default behavior.