power: If the lid is closed and we don't suspend, try to suspend again after a little while
Consider this situation:

1. An external monitor is plugged in.
2. You close the laptop's lid, expecting the laptop to suspend.
3. You unplug the monitor.  This isn't detected as it's a VGA connection.
4. You put the laptop in your backpack.
5. Your laptop melts, taking the chocolate bars in your backpack along with it.

The problem is when we don't suspend because an external monitor is still plugged in,
and we can't detect the unplug event (e.g. happens for VGA).

So, we install a 30-second timer when we *don't* suspend due to that situation.
When the timer expires, we'll check the monitor outputs again.  If there are no
monitors plugged in, we'll suspend as usual.