wacom: Avoid a warning at login
The wacom plugin calls init_screens before creating the devices
hash table in gsd_wacom_manager_idle_cb. init_screens connects to
the ::changed signal on GnomeRRScreens, and the signal handler
assumes that the devices hash table has been created. When we
are logging in, the ::changed signal is emitted before the idle
callback has a chance to run, causing a critical warning out
of g_hash_table_get_values.

Avoid that by ignoring the ::changed signal if the devices hash
table is not created yet.

https://bugzilla.gnome.org/show_bug.cgi?id=680291