power: fix idle blank and sleep timeouts
The previous logic for blank and sleep timeouts was broken, as it
setup the timeouts only after the timeout for brightness dimming
(which is logically unrelated). Refactor it to use GpmIdletime
instead of GTimeout for blank and sleep (so that actual idle times
are considered), and to take in account screensaver fading before
blanking.
The net effect of this patch is that the idle-dim-time has now
no effect on blank or sleep, and that setting session idleness to the
same time as sleep-display-* causes the screen to power off at
the end of the screensaver animation (which is the desired behaviour)

https://bugzilla.gnome.org/show_bug.cgi?id=668703