keyboard: Fix gtk+ IM module getting set to ibus for xkb sources
Instead of waiting until the async ibus engine set call to finish to
set the gtk+ IM module to 'ibus' we can just set the module
unconditionally. The user visible effect is the same, i.e. the user
will have to change to another input source because the current one
isn't working right and it simplifies the code.

https://bugzilla.gnome.org/show_bug.cgi?id=684037