keygrab: Allow unmodified and synchronous grabs if requested
Unmodified grabs are useful not only for a certain range of keysyms
but calling code has to deal with them specifically so make it
available on explicit request.

Likewise, synchronous grabs are useful as they allow us to replay
events that might be interesting to other applications but also need
special care on the calling side.

This patch changes the API so that both of the above options can be
requested and makes ungrabbing separate from the calling code
perspective.

https://bugzilla.gnome.org/show_bug.cgi?id=681685