xsettings: Optimize xsettings changes
I've noticed that we are updating XSettings several times during
login, which wakes up all running GTK+ apps, and possibly makes then
do a lot of work due to theme or font changes.

This patch addresses this in two ways:
1) Assume that gnome-shell will be running, and initialize the
   appmenu setting accordingly. This avoids a change when the shell
   comes on the bus later (g-s-d starts before the shell)
2) Batch updates in an idle

With these changes, I see XSettings getting set exactly once during
login, which is as it should be.

https://bugzilla.gnome.org/show_bug.cgi?id=686505